Create sp for each e2e run

+++ pkg/util/cluster/cluster.go
@@ -42,6 +42,7 @@ import (
 	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/features"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/network"
 	redhatopenshift20231122 "github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/redhatopenshift/2023-11-22/redhatopenshift"
+	"github.com/Azure/ARO-RP/pkg/util/azureerrors"
 	utilgraph "github.com/Azure/ARO-RP/pkg/util/graph"
 	"github.com/Azure/ARO-RP/pkg/util/rbac"
 	"github.com/Azure/ARO-RP/pkg/util/rolesets"
@@ -133,31 +134,26 @@ func New(log *logrus.Entry, environment env.Core, ci bool) (*Cluster, error) {
 	return c, nil
 }
 
-func (c *Cluster) CreateApp(ctx context.Context, clusterName string) error {
-	c.log.Infof("creating AAD application")
-	appID, appSecret, err := c.createApplication(ctx, "aro-"+clusterName)
-	if err != nil {
-		return err
-	}
-
-	c.log.Infof("creating service principal")
-	spID, err := c.createServicePrincipal(ctx, appID)
-	if err != nil {
-		return err
-	}
-
-	return os.WriteFile("clusterapp.env", []byte(fmt.Sprintf("export AZURE_CLUSTER_SERVICE_PRINCIPAL_ID=%s\nexport AZURE_CLUSTER_APP_ID=%s\nexport AZURE_CLUSTER_APP_SECRET=%s", spID, appID, appSecret)), 0o600)
+type appDetails struct {
+	applicationId     string
+	applicationSecret string
+	SPId              string
 }
 
-func (c *Cluster) DeleteApp(ctx context.Context) error {
-	err := env.ValidateVars(
-		"AZURE_CLUSTER_APP_ID",
-	)
+func (c *Cluster) createApp(ctx context.Context, clusterName string) (applicationDetails appDetails, err error) {
+	c.log.Infof("Creating AAD application")
+	appID, appSecret, err := c.createApplication(ctx, "aro-"+clusterName)
 	if err != nil {
-		return err
+		return appDetails{}, err
 	}
 
-	return c.deleteApplication(ctx, os.Getenv("AZURE_CLUSTER_APP_ID"))
+	c.log.Infof("Creating service principal")
+	spID, err := c.createServicePrincipal(ctx, appID)
+	if err != nil {
+		return appDetails{}, err
+	}
+
+	return appDetails{appID, appSecret, spID}, nil
 }
 
 func (c *Cluster) Create(ctx context.Context, vnetResourceGroup, clusterName string, osClusterVersion string) error {
@@ -170,20 +166,15 @@ func (c *Cluster) Create(ctx context.Context, vnetResourceGroup, clusterName str
 		return nil
 	}
 
-	err = env.ValidateVars(
-		"AZURE_FP_SERVICE_PRINCIPAL_ID",
-		"AZURE_CLUSTER_SERVICE_PRINCIPAL_ID",
-		"AZURE_CLUSTER_APP_ID",
-		"AZURE_CLUSTER_APP_SECRET",
-	)
+	fpSPId := os.Getenv("AZURE_FP_SERVICE_PRINCIPAL_ID")
+	if fpSPId == "" {
+		return fmt.Errorf("fp service principal id is not found")
+	}
+
+	appDetails, err := c.createApp(ctx, clusterName)
 	if err != nil {
 		return err
 	}
-
-	fpSPID := os.Getenv("AZURE_FP_SERVICE_PRINCIPAL_ID")
-	spID := os.Getenv("AZURE_CLUSTER_SERVICE_PRINCIPAL_ID")
-	appID := os.Getenv("AZURE_CLUSTER_APP_ID")
-	appSecret := os.Getenv("AZURE_CLUSTER_APP_SECRET")
 
 	visibility := api.VisibilityPublic
 
@@ -240,8 +231,8 @@ func (c *Cluster) Create(ctx context.Context, vnetResourceGroup, clusterName str
 	parameters := map[string]*arm.ParametersParameter{
 		"clusterName":               {Value: clusterName},
 		"ci":                        {Value: c.ci},
-		"clusterServicePrincipalId": {Value: spID},
-		"fpServicePrincipalId":      {Value: fpSPID},
+		"clusterServicePrincipalId": {Value: appDetails.SPId},
+		"fpServicePrincipalId":      {Value: fpSPId},
 		"vnetAddressPrefix":         {Value: addressPrefix},
 		"masterAddressPrefix":       {Value: masterSubnet},
 		"workerAddressPrefix":       {Value: workerSubnet},
@@ -292,7 +283,7 @@ func (c *Cluster) Create(ctx context.Context, vnetResourceGroup, clusterName str
 		{"/subscriptions/" + c.env.SubscriptionID() + "/resourceGroups/" + vnetResourceGroup + "/providers/Microsoft.Network/routeTables/" + clusterName + "-rt", rbac.RoleNetworkContributor},
 		{diskEncryptionSetID, rbac.RoleReader},
 	} {
-		for _, principalID := range []string{spID, fpSPID} {
+		for _, principalID := range []string{appDetails.SPId, fpSPId} {
 			for i := 0; i < 5; i++ {
 				_, err = c.roleassignments.Create(
 					ctx,
@@ -331,7 +322,7 @@ func (c *Cluster) Create(ctx context.Context, vnetResourceGroup, clusterName str
 	}
 
 	c.log.Info("creating cluster")
-	err = c.createCluster(ctx, vnetResourceGroup, clusterName, appID, appSecret, diskEncryptionSetID, visibility, osClusterVersion)
+	err = c.createCluster(ctx, vnetResourceGroup, clusterName, appDetails.applicationId, appDetails.applicationSecret, diskEncryptionSetID, visibility, osClusterVersion)
 
 	if err != nil {
 		return err
@@ -376,18 +367,29 @@ func (c *Cluster) generateSubnets() (vnetPrefix string, masterSubnet string, wor
 }
 
 func (c *Cluster) Delete(ctx context.Context, vnetResourceGroup, clusterName string) error {
+	c.log.Infof("Deleting cluster %s in resource group %s", clusterName, vnetResourceGroup)
 	var errs []error
 
-	switch {
-	case c.ci && env.IsLocalDevelopmentMode(): // PR E2E
+	if c.ci {
+		oc, err := c.openshiftclusters.Get(ctx, vnetResourceGroup, clusterName)
+		clusterResourceGroup := fmt.Sprintf("aro-%s", clusterName)
+		if err != nil {
+			c.log.Errorf("CI E2E cluster %s not found in resource group %s", clusterName, vnetResourceGroup)
+			errs = append(errs, err)
+		}
 		errs = append(errs,
+			c.deleteApplication(ctx, *oc.OpenShiftClusterProperties.ServicePrincipalProfile.ClientID),
 			c.deleteCluster(ctx, vnetResourceGroup, clusterName),
-			c.deleteClusterResourceGroup(ctx, vnetResourceGroup),
-			c.deleteVnetPeerings(ctx, vnetResourceGroup),
+			c.ensureResourceGroupDeleted(ctx, clusterResourceGroup),
+			c.deleteResourceGroup(ctx, vnetResourceGroup),
 		)
-	case c.ci: // Prod E2E
-		errs = append(errs, c.deleteClusterResourceGroup(ctx, vnetResourceGroup))
-	default:
+
+		if env.IsLocalDevelopmentMode() { //PR E2E
+			errs = append(errs,
+				c.deleteVnetPeerings(ctx, vnetResourceGroup),
+			)
+		}
+	} else {
 		errs = append(errs,
 			c.deleteRoleAssignments(ctx, vnetResourceGroup, clusterName),
 			c.deleteCluster(ctx, vnetResourceGroup, clusterName),
@@ -714,13 +716,28 @@ func (c *Cluster) deleteCluster(ctx context.Context, resourceGroup, clusterName 
 	return nil
 }
 
-func (c *Cluster) deleteClusterResourceGroup(ctx context.Context, resourceGroup string) error {
+func (c *Cluster) ensureResourceGroupDeleted(ctx context.Context, resourceGroupName string) error {
+	c.log.Printf("deleting resource group %s", resourceGroupName)
+	timeoutCtx, cancel := context.WithTimeout(ctx, 10*time.Minute)
+	defer cancel()
+
+	return wait.PollImmediateUntil(5*time.Second, func() (bool, error) {
+		_, err := c.groups.Get(ctx, resourceGroupName)
+		if azureerrors.ResourceGroupNotFound(err) {
+			c.log.Infof("finished deleting resource group %s", resourceGroupName)
+			return true, nil
+		}
+		return false, fmt.Errorf("failed to delete resource group %s with %s", resourceGroupName, err)
+	}, timeoutCtx.Done())
+}
+
+func (c *Cluster) deleteResourceGroup(ctx context.Context, resourceGroup string) error {
+	c.log.Printf("deleting resource group %s", resourceGroup)
 	if _, err := c.groups.Get(ctx, resourceGroup); err != nil {
 		c.log.Printf("error getting resource group %s, skipping deletion: %v", resourceGroup, err)
 		return nil
 	}
 
-	c.log.Printf("deleting resource group %s", resourceGroup)
 	if err := c.groups.DeleteAndWait(ctx, resourceGroup); err != nil {
 		return fmt.Errorf("error deleting resource group: %w", err)
 	}
